+++ examples/storybook/stories/Ricos/index.ts
@@ -1,6 +1,7 @@
 export { default as basicUsage } from './RicosBasicUsageStory';
 export { default as ExampleApp } from './ExampleAppStory';
 export { default as Tiptap } from './TiptapAppStory';
+export { default as TiptapInsideContainer } from './TiptapAppStoryWithContainer';
 export { default as staticToolbar } from './RicosStaticToolbar';
 export { default as modalAPI } from './RicosModalSettings';
 export { default as ricosMedia } from './RicosMediaStory';